framework + Рефакторинг
