+++ package.json
@@ -18,14 +18,14 @@
   "devDependencies": {
     "@babel/core": "7.17.10",
     "@babel/preset-env": "7.17.10",
-    "babel-loader": "8.2.4",
-    "copy-webpack-plugin": "10.2.4",
-    "css-loader": "^6.7.1",
+    "babel-loader": "^8.2.5",
+    "copy-webpack-plugin": "^10.2.4",
+    "css-loader": "6.7.1",
     "eslint": "8.13.0",
     "eslint-config-htmlacademy": "5.1.2",
-    "style-loader": "^3.3.1",
-    "webpack": "5.72.0",
-    "webpack-cli": "4.9.2",
+    "style-loader": "3.3.1",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
     "webpack-dev-server": "4.8.1"
   },
   "engines": {
