+++ src/model/point-model.js
@@ -1,7 +1,7 @@
 import { generatePoint } from '../mock/point.js';
 
 export default class PointModel {
-  #points = Array.from({length: 0}, generatePoint);
+  #points = Array.from({length: 10}, generatePoint);
 
   get points() {
     return this.#points;
