+++ src/utils/common.js
@@ -1,0 +1,44 @@
+const getRandomInt = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+const getRandomElement = (arr) => arr[Math.floor(Math.random() * arr.length)];
+
+const shuffleArr = (arr) => {
+  const newArr = arr.slice();
+
+  for (let i = arr.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [newArr[i], newArr[j]] = [newArr[j], newArr[i]];
+  }
+
+  return newArr;
+};
+
+const getRandomArr = (arr) => {
+  const randomArr = [];
+
+  if (arr.length !== 0) {
+    const count = getRandomInt(1, arr.length);
+    const elements = shuffleArr(arr);
+
+    for (let i = 0; i < count; i++) {
+      randomArr.push(elements[i]);
+    }
+  }
+
+  return randomArr;
+};
+
+const capitalizeFirstLetter = (str) => (str) ? str[0].toUpperCase() + str.slice(1) : str;
+
+export {
+  getRandomInt,
+  getRandomElement,
+  shuffleArr,
+  getRandomArr,
+  capitalizeFirstLetter
+};
